Find icons for file candidates containing a path prefix.
Certain functions (e.g., `project-find-files') will produce a
candidate list where each candidate may contain a path prefix (e.g.,
"docs/Makefile").  This is problematic when invoking
`all-the-icons-icon-for-file' as it is expecting a filename, not a
path.  Some regular expressions in `all-the-icons-regexp-icon-alist'
expect to match against the entire filename (e.g., "^Makefile$") and
will fail to match when a path prefix is present.  This change strips
any path prefix before invoking `all-the-icons-icon-for-file'.